requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
--trusted-host pypi.python.org
--trusted-host pypi.org
--trusted-host files.pythonhosted.org

asgiref==3.7.2
    # via django
django==4.2.8
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
environs==9.5.0
    # via -r requirements.in
marshmallow==3.20.1
    # via environs
packaging==23.2
    # via marshmallow
pillow==10.1.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via environs
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
